[MissionsFilters] create filtersFunctions
