+++ frontend/src/domain/use_cases/missions/filters/unitFilterFunction.ts
@@ -1,0 +1,15 @@
+import type { MissionType } from '../../../entities/missions'
+
+export function unitFilterFunction(mission: MissionType, filter: string[]) {
+  if (filter.length > 0) {
+    return !!mission.controlUnits.find(controlUnit => {
+      if (filter.find(unit => unit === controlUnit.name)) {
+        return controlUnit
+      }
+
+      return undefined
+    })
+  }
+
+  return mission
+}
